api-gateway/gateway/requirement.txt
Django==4.2.2
djangorestframework==3.14.0
django-cors-headers==4.0.0
PyJWT==2.7.0
python-eureka-client==0.11.7
requests==2.31.0
